Create folder setup and moved classes
