--- src/main/java/org/example/fishtank/infrastructure/persistence/UserRepository.java
@@ -1,8 +1,0 @@
-package org.example.fishtank.infrastructure.persistence;
-
-import org.example.fishtank.domain.entity.AppUser;
-import org.example.fishtank.domain.entity.Fish;
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository <AppUser, Integer> {
-}
