+++ src/main/java/org/example/fishtank/model/entity/AppUser.java
@@ -1,4 +1,4 @@
-package org.example.fishtank.domain.entity;
+package org.example.fishtank.model.entity;
 
 import jakarta.persistence.*;
