+++ src/main/java/org/example/fishtank/repository/FishRepository.java
@@ -1,11 +1,10 @@
-package org.example.fishtank.infrastructure.persistence;
+package org.example.fishtank.repository;
 
-import org.example.fishtank.domain.entity.Fish;
+import org.example.fishtank.model.entity.Fish;
 import org.springframework.data.repository.ListCrudRepository;
 
 
 
 public interface FishRepository extends ListCrudRepository<Fish, Integer> {
-
 
 }
