+++ src/main/java/org/example/fishtank/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package org.example.fishtank.repository;
+
+import org.example.fishtank.model.entity.AppUser;
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserRepository extends CrudRepository <AppUser, Integer> {
+}
